Refactor Makefile
- Use variables to organize prerequisites
- Use less phony targets
- Rename files
- Suggest using `npm install` instead of `yarn`
