+++ Makefile
@@ -1,28 +1,44 @@
 BIN = $(PWD)/node_modules/.bin
+
 LEZER_ARGS :=
 ifdef DEBUG
 	LEZER_ARGS := --names $(LEZER_ARGS)
 endif
 
+# Written by humans
+SRC = src/julia.grammar src/tokens.js
+
+# Generated by lezer
+PARSER = src/_parser.js src/_parser.terms.js
+
+# Bundled by rollup
+INDEX = dist/index.cjs dist/index.es.js
+
+
 .DELETE_ON_ERROR:
 
-.PHONY: default
-default: build
+build: $(INDEX)
 
-.PHONY: build
-build: dist/index.cjs dist/index.es.js
+$(INDEX): $(PARSER)
+	ROLLUP_IN="$(<)" \
+	ROLLUP_OUT="dist/index" \
+	$(BIN)/rollup -c
 
-.PHONY: test
-test:
+# If the parser generation fails, you might need to set the following environment variable:
+# export NODE_OPTIONS="--max-old-space-size=8192"
+$(PARSER): $(SRC)
+	$(BIN)/lezer-generator $(LEZER_ARGS) $(<) -o src/_parser
+
+
+.PHONY: test check clean
+
+test: $(PARSER) test/*
 	$(BIN)/mocha test/test-julia.js
 
-.PRECIOUS: src/%.js src/%.terms.js
-src/%.js src/%.terms.js: src/%.grammar
-	@echo 'Compiling $(<)'
-	$(BIN)/lezer-generator $(LEZER_ARGS) $< -o $(<:%.grammar=%)
+check: # Ensure tooling is installed
+	node --version
+	npm --version
 
-dist/%.cjs dist/%.es.js: src/%.js src/%.tokens.js src/%.terms.js
-	@echo 'Bunding $(<)'
-	ROLLUP_IN="$(<)" \
-	ROLLUP_OUT="$(<:src/%.js=dist/%)" \
-	$(BIN)/rollup -c
+clean: # Remove generated files
+	rm $(PARSER)
+	rm $(INDEX)
