+++ src/tokens.js
@@ -1,5 +1,5 @@
 import { ExternalTokenizer } from "@lezer/lr";
-import * as terms from "./index.terms.js";
+import * as terms from "./_parser.terms.js";
 
 // UNICODE CODEPOINTS
